+++ packages/atomic/src/components/atomic-query-summary/atomic-query-summary.stories.tsx
@@ -1,9 +1,0 @@
-import defaultStory from '../../../.storybook/default-story';
-
-const {defaultModuleExport, exportedStory} = defaultStory(
-  'Atomic/QuerySummary',
-  'atomic-query-summary',
-  {}
-);
-export default defaultModuleExport;
-export const DefaultQuerySummary = exportedStory;