+++ packages/atomic/src/components/search/atomic-no-results/atomic-no-results.stories.tsx
@@ -1,0 +1,20 @@
+import defaultStory from '../../../../.storybook/default-story';
+
+const {defaultModuleExport, exportedStory} = defaultStory(
+  'Atomic/NoResults',
+  'atomic-no-results',
+  {},
+  {
+    engineConfig: {
+      search: {
+        preprocessSearchResponseMiddleware: (r) => {
+          r.body.results = [];
+          return r;
+        },
+      },
+    },
+  }
+);
+
+export default defaultModuleExport;
+export const DefaultNoResults = exportedStory;