+++ packages/auth/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.7](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.6...@coveo/auth@1.6.7) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/auth
+
+
+
+
+
+## [1.6.6](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.5...@coveo/auth@1.6.6) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/auth
+
+
+
+
+
+## [1.6.5](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.4...@coveo/auth@1.6.5) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/auth
+
+
+
+
+
+## [1.6.4](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.3...@coveo/auth@1.6.4) (2022-06-16)
+
+**Note:** Version bump only for package @coveo/auth
+
+
+
+
+
+## [1.6.3](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.1...@coveo/auth@1.6.3) (2022-06-16)
+
+
+### Bug Fixes
+
+* commit package-lock files right when bumping, undo last bump & fix region ([#2074](https://github.com/coveo/ui-kit/issues/2074)) ([28ad872](https://github.com/coveo/ui-kit/commit/28ad872363d2b84711865cda0c6d61bba7354927))
+
+
+
+
+
 ## [1.6.2](https://github.com/coveo/ui-kit/compare/@coveo/auth@1.6.1...@coveo/auth@1.6.2) (2022-06-13)