+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.25.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.4...@coveo/quantic@1.25.5) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.3...@coveo/quantic@1.25.4) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.2...@coveo/quantic@1.25.3) (2022-06-17)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.2](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.1...@coveo/quantic@1.25.2) (2022-06-16)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.25.1](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.25.0...@coveo/quantic@1.25.1) (2022-06-16)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 # [1.25.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.24.0...@coveo/quantic@1.25.0) (2022-06-13)